+++ src/Label.ts
@@ -1,0 +1,4 @@
+export class Label {
+  public constructor(public readonly name: string) {}
+}
+