+++ src/index.ts
@@ -1,1 +1,2 @@
 export { Step } from './Step'
+export { Label } from './Label'