+++ test/Step.test.ts
@@ -1,4 +1,5 @@
 import * as sedk from '../src'
+import { database } from './database'
 
 describe('simple Cypher', () => {
   const cypher = new sedk.Step()
@@ -6,12 +7,13 @@ describe('simple Cypher', () => {
     cypher.cleanUp()
   })
   it('Produce: [MATCH (n:Person) RETURN n]', () => {
-    const actual = cypher.match(['Person']).return().getCypher()
+    const actual = cypher.match([database.Labels.Person]).return().getCypher()
 
     expect(actual).toBe('MATCH (n:Person) RETURN n')
   })
   it('Produce: [MATCH (n:Person:Animal) RETURN n]', () => {
-    const actual = cypher.match(['Person', 'Animal']).return().getCypher()
+    const labels = [database.Labels.Person, database.Labels.Animal]
+    const actual = cypher.match(labels).return().getCypher()
 
     expect(actual).toBe('MATCH (n:Person:Animal) RETURN n')
   })