+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [0.0.2]  2022-07-17
 ### Added
 - documentation
+- make Label a class
 
 ## [0.0.1]  2022-07-17
 ### Added